+++ io/usart.hpp
@@ -11,9 +11,9 @@ public:
 
 	}
 	bool Init(USART_TypeDef * device, uint32_t baud);
-	void Send(const uint8_t* data, int16_t size);
-	void Send(const char* data, int16_t size) { Send((const uint8_t*)data, size); }
-	int16_t Read(uint8_t* data, int16_t max_size);
+	void Send(const uint8_t* data, int32_t size);
+	void Send(const char* data, int32_t size) { Send((const uint8_t*)data, size); }
+	int32_t Read(uint8_t* data, int32_t max_size);
 
 	bool HasData();
 	void handleIRQ();
@@ -25,13 +25,13 @@ private:
 	USART_TypeDef * device_;
 
 	uint8_t txBuffer[TX_BUFFER_SIZE];
-	int16_t txBufferReadIdx = 0;
-	int16_t txBufferWriteIdx = 0;
+	int32_t txBufferReadIdx = 0;
+	int32_t txBufferWriteIdx = 0;
 	bool txStarted = 0;
 
 	uint8_t rxBuffer[RX_BUFFER_SIZE];
-	int16_t rxBufferReadIdx = 0;
-	int16_t rxBufferWriteIdx = 0;
+	int32_t rxBufferReadIdx = 0;
+	int32_t rxBufferWriteIdx = 0;
 	bool rxEmpty = 0;
 
 	DISALLOW_COPY_AND_ASSIGN(Usart);
